выполняет дз по модуляции
